requirements.txt
numpy>=2.4.2
pygame-ce>=2.5.6
